server: Fix a pointer to uint64_t cast in spice_replay_next_cmd()
This avoids a compilation error with -Werror on 32 bit systems as the
pointer size differs from that of an uint64_t.

+++ server/red-replay-qxl.c
@@ -1156,7 +1156,7 @@ SPICE_GNUC_VISIBLE QXLCommandExt* spice_replay_next_cmd(SpiceReplay *replay,
     case QXL_CMD_UPDATE:
     case QXL_CMD_SURFACE:
         info = (QXLReleaseInfo *)cmd->cmd.data;
-        info->id = (uint64_t)cmd;
+        info->id = (uint64_t)(uintptr_t)cmd;
     }
 
     replay->counter++;
